+++ src/main/java/dev/lambdaurora/lambdynlights/LambDynLightsConstants.java
@@ -15,7 +15,7 @@ import net.fabricmc.loader.api.FabricLoader;
  * Contains constants about LambDynamicLights.
  *
  * @author LambdAurora
- * @version 3.0.1
+ * @version 3.1.2
  * @since 3.0.1
  */
 public final class LambDynLightsConstants {
@@ -23,6 +23,11 @@ public final class LambDynLightsConstants {
 	 * The namespace of this mod, whose value is {@value}.
 	 */
 	public static final String NAMESPACE = "lambdynlights";
+
+	/**
+	 * The unsupported development mode text.
+	 */
+	public static final String DEV_MODE_OVERLAY_TEXT = "[LambDynamicLights Dev Version (Unsupported)]";
 
 	/**
 	 * {@return {@code true} if this mod is in development mode, or {@code false} otherwise}