+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/DebugScreenOverlayMixin.java
@@ -10,6 +10,7 @@
 package dev.lambdaurora.lambdynlights.mixin;
 
 import dev.lambdaurora.lambdynlights.LambDynLights;
+import dev.lambdaurora.lambdynlights.LambDynLightsConstants;
 import net.minecraft.TextFormatting;
 import net.minecraft.client.gui.components.DebugScreenOverlay;
 import org.spongepowered.asm.mixin.Mixin;
@@ -23,7 +24,7 @@ import java.util.List;
  * Adds a debug string for dynamic light sources tracking and updates.
  *
  * @author LambdAurora
- * @version 1.3.2
+ * @version 3.1.2
  * @since 1.3.2
  */
 @Mixin(DebugScreenOverlay.class)
@@ -46,5 +47,9 @@ public class DebugScreenOverlayMixin {
 
 		builder.append(')');
 		list.add(builder.toString());
+
+		if (LambDynLightsConstants.isDevMode()) {
+			list.add(TextFormatting.RED + LambDynLightsConstants.DEV_MODE_OVERLAY_TEXT);
+		}
 	}
 }